refactor(justice): extract shared ODS helpers into _base
DoJ NI bulletins all ship as accessibility-format ODS workbooks with the
same footnote markers, suppression codes and cell-repetition quirks. Move
the parsing primitives out of nicts_quarterly so the next DoJ module can
reuse them rather than duplicating ~55 lines.
