+++ src/bolster/data_sources/justice/nicts_quarterly.py
@@ -57,11 +57,12 @@ from urllib.parse import unquote, urljoin
 import bs4
 import pandas as pd
 from odf.opendocument import load as load_ods
-from odf.table import Table, TableCell, TableRow
-from odf.text import P
+from odf.table import Table
 
 from bolster.utils.cache import CachedDownloader, DownloadError
 from bolster.utils.web import session
+
+from ._base import _parse_value, _sheet_rows, _strip_note_refs
 
 logger = logging.getLogger(__name__)
 
@@ -77,18 +78,6 @@ _PERIOD_RE = re.compile(r"^(\d{4})\s+(Total|Q[1-4])$")
 # Bulletin filenames encode their reporting quarter as a month range
 _FILENAME_PERIOD_RE = re.compile(r"(january|april|july|october)\s*to\s*(march|june|september|december)\D{0,4}(\d{4})")
 _MONTH_TO_QUARTER = {"january": 1, "april": 2, "july": 3, "october": 4}
-
-# Suppression markers used in place of a value
-_SUPPRESSED = {"", "-", "N/A", "n/a", "[z]", "[c]", "[x]", ":"}
-
-# Durations are published as HH:MM and can exceed 24 hours
-_DURATION_RE = re.compile(r"^(\d+):([0-5]\d)$")
-
-# Footnote references embedded in labels, e.g. "Total [note 3]"
-_NOTE_REF_RE = re.compile(r"\s*[\[{]note \d+[\]}]", re.IGNORECASE)
-
-# Cell repetition attributes can run to thousands for trailing filler
-_MAX_REPEAT = 60
 
 # Bulletins are provisional but only republished quarterly
 _CACHE_TTL_HOURS = 24 * 90
@@ -136,52 +125,6 @@ def _normalise_sheet_name(name: str) -> str:
     return re.sub(r"_+", "_", re.sub(r"[^a-z0-9]+", "_", name.lower())).strip("_")
 
 
-def _strip_note_refs(text: str) -> str:
-    """Remove ``[note N]`` footnote markers from a label.
-
-    Example:
-        >>> _strip_note_refs("Total [note 3] [note 32]")
-        'Total'
-    """
-    return _NOTE_REF_RE.sub("", text).strip()
-
-
-def _parse_value(text: str) -> float:
-    """Convert a raw cell string into a numeric value.
-
-    Handles the four shapes NICTS publishes: comma-grouped integers, plain
-    decimals, ``HH:MM`` durations (converted to fractional hours, and which may
-    exceed 24 hours), and suppression markers such as ``[z]`` or ``N/A``.
-
-    Args:
-        text: Raw cell text.
-
-    Returns:
-        The numeric value, or NaN when suppressed or unparseable.
-
-    Example:
-        >>> _parse_value("1,234")
-        1234.0
-        >>> _parse_value("6485:30")
-        6485.5
-        >>> import math
-        >>> math.isnan(_parse_value("[z]"))
-        True
-    """
-    text = text.strip()
-    if text in _SUPPRESSED:
-        return float("nan")
-
-    duration = _DURATION_RE.match(text)
-    if duration:
-        return int(duration.group(1)) + int(duration.group(2)) / 60
-
-    try:
-        return float(text.replace(",", ""))
-    except ValueError:
-        return float("nan")
-
-
 def _parse_period(label: str) -> tuple[int, int | None] | None:
     """Parse a period column header into (year, quarter).
 
@@ -206,36 +149,6 @@ def _parse_period(label: str) -> tuple[int, int | None] | None:
     year = int(match.group(1))
     marker = match.group(2)
     return year, None if marker == "Total" else int(marker[1])
-
-
-def _sheet_rows(table: Table) -> list[list[str]]:
-    """Extract a worksheet as a list of raw text rows.
-
-    pandas' ODF reader coerces time-typed cells via :class:`pandas.Timestamp`,
-    which fails on NICTS sitting-time durations over 24 hours. Reading the
-    underlying cell text directly sidesteps that entirely.
-
-    Args:
-        table: ODF table element.
-
-    Returns:
-        List of rows, each a list of cell strings, with trailing blanks removed.
-    """
-    rows: list[list[str]] = []
-    for row in table.getElementsByType(TableRow):
-        row_repeat = min(int(row.getAttribute("numberrowsrepeated") or 1), _MAX_REPEAT)
-        cells: list[str] = []
-        for cell in row.getElementsByType(TableCell):
-            col_repeat = min(int(cell.getAttribute("numbercolumnsrepeated") or 1), _MAX_REPEAT)
-            text = " ".join(str(p) for p in cell.getElementsByType(P))
-            cells.extend([text] * col_repeat)
-        while cells and not cells[-1].strip():
-            cells.pop()
-        if not cells:
-            continue
-        for _ in range(row_repeat):
-            rows.append(list(cells))
-    return rows
 
 
 def _parse_sheet(court: str, rows: list[list[str]]) -> pd.DataFrame:
